+++ content-learn/mastering-meshery-istio/index.mdx
@@ -1,6 +1,6 @@
 ---
-title: 'Mastering Meshery and Istio'
-description: 'Learn how to build your own addons that will super charge your development'
+title: 'Mastering Service Mesh'
+description: 'Learn how to setup a service mesh and build an application'
 heroDescription: 'Addons give you superpowers to extend Storybook, automate workflows, and integrate with your favorite tools. This guide shows you how to create an addon.'
 overview: 'While the Storybook community offers over 250 addons, you can also build one tailored to your specific needs. This guide introduces you to the Addon Kit and APIs, along the way you will build an addon from scratch.'
 order: 4
@@ -23,6 +23,7 @@ contributors: []
 twitterShareText: "I’m learning how to build a Storybook addon! They're great for customizing Storybook to fit your UI development workflow."
 ---
 
-<h2>What you'll build</h2>
-
-Mastering service mesh learning path.
+<!-- 
+  This file is only used to render the courses list within a learning path.
+  Check the Learn-Layer5 folder under src/sections/, src/templates for more understanding of how the data is used
+-->
