+++ content-learn/mastering-meshery-istio/introduction-to-service-meshes/index.mdx
@@ -1,11 +1,10 @@
 ---
 docType: "Course"
 title: "Course"
-tocTitle: "Introduction"
 description: "Meshery is the service mesh managemen plane which offers lifecycle, configuration, and performance management of service meshes and their workloads."
 videos: 4
 lectures: 12
-courseTitle: "Securing The Service Mesh"
+courseTitle: "Introduction to Service Meshes - Hands On"
 themeColor: "#00B39F"
 heroAnimationName: null
 cardImage: "../../../src/assets/images/service-mesh-icons/kuma-white.svg"
@@ -34,4 +33,7 @@ meshesYouLearn:
   ]
 ---
 
-Chapters available
+<!-- 
+  This file is only used to render the course overview page for the respective course within a learn path.
+  Check the Learn-Layer5 folder under src/sections/, src/templates for more understanding of how the data is used
+-->
